Redraw the logo (although only minor appearance changes).
Previously, the borders had distinct objects, and certain things were
misaligned. The "T" wasn't in center properly, due to a small quirk in
the right top of the path.

This commit introduces a hand-written logo SVG, which uses only 3
objects (!) and uses more round numbers for aesthetics.

so, basically, nothing changed

i need a life