Initial range coder for zmicro, a compression algorithm that works well on small inputs.
We introduce the first component in zmicro, which is intended as default compression algorithm since we have decreased the size of clusters. Other algorithms are generally not sufficient as they store some header or require a lot of input to inhabitate the dictionary. Zmicro bases on adaptive range coding, with techniques predicting the next bit and using the calculated probability to minimize the space usage.