TFS is now an async file system (through futures)
This commit changes how I/O is handled entirely. It moves from a
syncrnized, blocking model to an async, non-blocking model.

To avoid callback hell, we rely on the `futures` crate (by @alexcrichton
et co.), which provides a nice way of handling operations and
completion.

One piece is still missing: an alternative to `Mutex` which is
non-blocking. I have already got the idea how to do this (lazy
queueing), but it is not implemented yet.