Allow overlapping memory reclamation cycles/garbage collection.
This commit removes the short-circuit happening when two threads are pending on availability of garbage collection. Instead, it makes the span in which this condition exists (i.e. while updating the snapshot stack) shorter, meaning that the blocking time will be relatively short, and futhermore allowing multiple threads to GC at the same time (for example if a destructor leaves a lot of garbage behind, a second GC can happen).

TODO: Use a more fine-grained, global metric for when to run a cycle of garbage collection