Reserve special pointers for `conc::hazard::State` to avoid overlapping with pointers used in `Protect`.
This commit adds a few statics, representing the special states of the hazard, thus avoiding problems like the ones partially fixed by ce10d585dab1ac26bc377c0cfbe37fedecaac189. This amounts to a complete fix of these issues, in the sense that you can use any value, as long as it doesn't use these objects, which would mean taking objects, you don't even control in the RO-data section and using them as trap-values, which is - if not UB - never going to happen in practice.