Prevent certain tests in `conc` from messing with each other by having seperated states.
This moves certain parts of the `global` API into `State` such that the
tests can create a state and manipulate this, fixing bugs where one test
would add panicking garbage, which would then be collected by another
test, thus messing up the tests. Here the tests with panicking garbage
simply create a state, add stuff to that, and move on.

- Test loose hazards when the process ends.
- Set #[should_panic] after #[test].